deployment/backend/generator/requirements.txt
fastapi
uvicorn
requests
python-docx 
mistune